refactor: implement dynamic workspace layout with persistent sidebar
- Create WorkspaceLayout component with persistent sidebar navigation
- Extract IssueDetails into reusable component from page
- Add WorkspaceContent component for dynamic view switching
- Update IssuesList to support client-side navigation
- Refactor workspace pages to use new layout system
- Eliminate full page refreshes when navigating between issues

The sidebar and header now remain persistent while only the content area updates, providing a much better user experience.
